projects/sports-news-packager-bot/config.py
"""
Переменные окружения для бота упаковки новостей Sports.ru.
"""
import os

from dotenv import load_dotenv

load_dotenv()

BOT_TOKEN = os.getenv("BOT_TOKEN")
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
# Бесплатный tier в AI Studio; при желании задайте GEMINI_MODEL=gemini-2.0-flash
GEMINI_MODEL = os.getenv("GEMINI_MODEL", "gemini-1.5-flash")
GEMINI_MAX_OUTPUT_TOKENS = int(os.getenv("GEMINI_MAX_OUTPUT_TOKENS", "8192"))
