projects/sports-news-packager-bot/requirements.txt
python-telegram-bot==21.8
python-dotenv==1.0.1
google-generativeai>=0.8.0
